Schemas can't be dropped if the migration table exists
Trying to drop a schema when the migrations table still exists (even
after a migration:reset) causes the error

cannot drop schema fugit_eveniet because other objects depend on it
DETAIL:  table migrations depends on schema fugit_eveniet
HINT:  Use DROP ... CASCADE to drop the dependent objects too.

The simplest solution to this is to simply add 'CASCADE' to the drop
query.